initial commit of mtclim removal and initialize_atmos refactor for classic driver
