+++ docs/Documentation/Drivers/Classic/DefaultOutputs.md
@@ -2,11 +2,6 @@
 
 If the user does not give explicit output file instructions in the global parameter file, then VIC will produce a set of default output files, one per grid cell. The exact set of output files, and which variables they contain, is determined by model settings, as follows:
 
-*   If OUTPUT_FORCE is TRUE in the [global parameter file](GlobalParam.md), VIC will produce 1 output file per grid cell, containing: PREC, AIR_TEMP, SHORTWAVE, LONGWAVE, DENSITY, PRESSURE, VP, and WIND. Units of these variables will depend on whether ALMA_OUTPUT is set to TRUE or FALSE in the [global parameter file](GlobalParam.md). These files will be named "full_data_lat_lon", where "lat" and "lon" are the latitude and longitude of the grid cell.
-*   If OUTPUT_FORCE is FALSE in the [global parameter file](GlobalParam.md), the set of output files, and the variables they contain, will depend on which operation mode VIC is being run in (e.g. FULL_ENERGY, FROZEN_SOIL, etc. in the [global parameter file](GlobalParam.md)).
-
-VIC can also be explicitly instructed to output these default files by pasting the contents of the file output.TRADITIONAL.410.template, which is included with the [VIC source code](../../../SourceCode/Code.md), into the bottom of your [global parameter file](GlobalParam.md). Of course, this is not necessary, since omitting any output file format information will cause VIC to make these files by default. But if you wish to have output that is similar to the default output files, but with fewer or additional variables, it is easiest to paste the contents into the global parameter file and then edit them to your liking. NOTE: to have output files exactly similar to those of 4.0.6 and earlier, use output.TRADITIONAL.template, instead of output.TRADITIONAL.410.template.
-
 ## Flux Files
 
 The primary output file type for the VIC-NL model is the flux file, which contains information about moisture and energy fluxes for each time step. These output files are based on the model output files used for the PILPS-2C project. The number of variables in this file depends on the values of FULL_ENERGY and FROZEN_SOIL in the [global parameter file](GlobalParam.md); when either FULL_ENERGY or FROZEN_SOIL are true, the file will contain the same number of columns as found in the PILPS-2C files for comparison purposes. Flux files are always written, regardless of the mode of operation. Flux files begin with the prefix "fluxes_".
