+++ vic/drivers/python/vic_headers.py
@@ -55,27 +55,6 @@ enum
 };
 enum
 {
-    VP_ITER_NONE,
-    VP_ITER_ALWAYS,
-    VP_ITER_ANNUAL,
-    VP_ITER_CONVERGE
-};
-enum
-{
-    LW_TVA,
-    LW_ANDERSON,
-    LW_BRUTSAERT,
-    LW_SATTERLUND,
-    LW_IDSO,
-    LW_PRATA
-};
-enum
-{
-    LW_CLOUD_BRAS,
-    LW_CLOUD_DEARDORFF
-};
-enum
-{
     FROM_VEGLIB,
     FROM_VEGPARAM
 };
@@ -121,8 +100,6 @@ enum
     REL_HUMID,
     SHORTWAVE,
     SNOWF,
-    TMAX,
-    TMIN,
     TSKC,
     VEGCOVER,
     VP,
@@ -399,9 +376,6 @@ typedef struct {
     _Bool IMPLICIT;
     _Bool JULY_TAVG_SUPPLIED;
     _Bool LAKES;
-    unsigned short int LW_CLOUD;
-    unsigned short int LW_TYPE;
-    _Bool MTCLIM_SWE_CORR;
     size_t Ncanopy;
     size_t Nfrost;
     size_t Nlakenode;
@@ -409,7 +383,6 @@ typedef struct {
     size_t Nnode;
     _Bool NOFLUX;
     size_t NVEGTYPES;
-    _Bool PLAPSE;
     unsigned short int RC_MODE;
     size_t ROOT_ZONES;
     _Bool QUICK_FLUX;
@@ -443,7 +416,6 @@ typedef struct {
     _Bool COMPRESS;
     _Bool MOISTFRACT;
     size_t Noutfiles;
-    _Bool OUTPUT_FORCE;
     _Bool PRT_HEADER;
     _Bool PRT_SNOW_BAND;
 } option_struct;
@@ -504,22 +476,6 @@ typedef struct {
     double CANOPY_CLOSURE;
     double CANOPY_RSMAX;
     double CANOPY_VPDMINFACTOR;
-    double MTCLIM_TDAYCOEF;
-    double MTCLIM_SOLAR_CONSTANT;
-    double MTCLIM_SNOW_TCRIT;
-    double MTCLIM_SNOW_TRATE;
-    double MTCLIM_TBASE;
-    double MTCLIM_ABASE;
-    double MTCLIM_C;
-    double MTCLIM_B0;
-    double MTCLIM_B1;
-    double MTCLIM_B2;
-    double MTCLIM_RAIN_SCALAR;
-    double MTCLIM_DIF_ALB;
-    double MTCLIM_SC_INT;
-    double MTCLIM_SC_SLOPE;
-    double MTCLIM_SRADDT;
-    double MTCLIM_SW_PREC_THRESH;
     double LAKE_TMELT;
     double LAKE_MAX_SURFACE;
     double LAKE_BETA;
