+++ vic/drivers/shared/src/get_parameters.c
@@ -124,55 +124,6 @@ get_parameters(FILE *paramfile)
             else if (strcasecmp("CANOPY_VPDMINFACTOR", optstr) == 0) {
                 sscanf(cmdstr, "%*s %lf", &param.CANOPY_VPDMINFACTOR);
             }
-            // MTCLIM Parameters
-            else if (strcasecmp("MTCLIM_SOLAR_CONSTANT", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_SOLAR_CONSTANT);
-            }
-            else if (strcasecmp("MTCLIM_TDAYCOEF", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_TDAYCOEF);
-            }
-            else if (strcasecmp("MTCLIM_SNOW_TCRIT", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_SNOW_TCRIT);
-            }
-            else if (strcasecmp("MTCLIM_SNOW_TRATE", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_SNOW_TRATE);
-            }
-            else if (strcasecmp("MTCLIM_TBASE", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_TBASE);
-            }
-            else if (strcasecmp("MTCLIM_ABASE", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_ABASE);
-            }
-            else if (strcasecmp("MTCLIM_C", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_C);
-            }
-            else if (strcasecmp("MTCLIM_B0", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_B0);
-            }
-            else if (strcasecmp("MTCLIM_B1", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_B1);
-            }
-            else if (strcasecmp("MTCLIM_B2", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_B2);
-            }
-            else if (strcasecmp("MTCLIM_RAIN_SCALAR", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_RAIN_SCALAR);
-            }
-            else if (strcasecmp("MTCLIM_DIF_ALB", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_DIF_ALB);
-            }
-            else if (strcasecmp("MTCLIM_SC_INT", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_SC_INT);
-            }
-            else if (strcasecmp("MTCLIM_SC_SLOPE", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_SC_SLOPE);
-            }
-            else if (strcasecmp("MTCLIM_SRADDT", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_SRADDT);
-            }
-            else if (strcasecmp("MTCLIM_SW_PREC_THRESH", optstr) == 0) {
-                sscanf(cmdstr, "%*s %lf", &param.MTCLIM_SW_PREC_THRESH);
-            }
             // Lake Parameters
             else if (strcasecmp("LAKE_TMELT", optstr) == 0) {
                 sscanf(cmdstr, "%*s %lf", &param.LAKE_TMELT);
@@ -649,39 +600,7 @@ validate_parameters()
         log_err(
             "CANOPY_VPDMINFACTOR must be defined on the interval [0, inf) (-)");
     }
-    // MTCLIM Parameters
-    if (!(param.MTCLIM_SOLAR_CONSTANT >= 0.)) {
-        log_err(
-            "MTCLIM_SNOW_TRATE must be defined on the interval [0, inf) (W/m2)");
-    }
-    // MTCLIM_TDAYCOEF - Currently, no constraints
-    // MTCLIM_SNOW_TCRIT - Currently, no constraints
-    // MTCLIM_SNOW_TRATE - Currently, no constraints
-    // MTCLIM_TBASE - Currently, no constraints
-    // MTCLIM_ABASE - Currently, no constraints
-    // MTCLIM_C - Currently, no constraints
-    // MTCLIM_B0 - Currently, no constraints
-    // MTCLIM_B1 - Currently, no constraints
-    // MTCLIM_B2 - Currently, no constraints
-    // MTCLIM_RAIN_SCALAR - Currently, no constraints
-    if (!(param.MTCLIM_DIF_ALB >= 0 && param.MTCLIM_DIF_ALB <= 1)) {
-        log_err("MTCLIM_DIF_ALB must be defined on the interval [0,1] (-)")
-    }
-    if (!(param.MTCLIM_DIF_ALB >= 0 && param.MTCLIM_DIF_ALB <= 1)) {
-        log_err("MTCLIM_DIF_ALB must be defined on the interval [0,1] (-)")
-    }
-    if (!(param.MTCLIM_SC_INT >= 0.)) {
-        log_err(
-            "MTCLIM_SC_INT must be defined on the interval [0, inf) (MJ/m2/day)");
-    }
-    // MTCLIM_SC_SLOPE - Currently, no constraints
-    if (!(param.MTCLIM_SRADDT >= 0 && param.MTCLIM_SRADDT <= SEC_PER_DAY)) {
-        log_err(
-            "MTCLIM_SRADDT must be defined on the interval [0, 86400) (seconds)");
-    }
-    if (!(param.MTCLIM_SW_PREC_THRESH >= 0.)) {
-        log_err("MTCLIM_SRADDT must be defined on the interval [0, inf) (mm)");
-    }
+
     // Lake Parameters
     // LAKE_TMELT - Currently, no constraints
     if (!(param.LAKE_MAX_SURFACE >= 0.)) {
